# backend/auth-facade/openapi.yaml
openapi: 3.1.0
info:
  title: Auth Facade API
  description: |
    Backend for Frontend (BFF) authentication service for the EMS platform.

    ## Overview
    The Auth Facade provides provider-agnostic authentication, abstracting identity providers
    (Keycloak, Auth0, Okta, Azure AD) behind a unified API. This allows switching identity
    providers via YAML configuration without code changes.

    ## Architecture
    - **Strategy Pattern**: Identity providers implement the `IdentityProvider` interface
    - **Provider Selection**: Set `auth.facade.provider` in configuration (keycloak, auth0, okta, azure-ad)
    - **Multi-tenancy**: Tenant ID passed via `X-Tenant-ID` header, maps to IdP realm/tenant

    ## Authentication Flows
    1. **Direct Login**: Email/password via Resource Owner Password Credentials
    2. **Social Login**: Google One Tap, Microsoft MSAL token exchange
    3. **Provider Redirect**: OAuth2/OIDC redirect via `kc_idp_hint` (dynamic provider selection)
    4. **MFA**: TOTP-based two-factor authentication

    ## Token Management
    - Access tokens: Short-lived JWT (5-15 minutes)
    - Refresh tokens: Longer-lived, with rotation on use
    - Token blacklisting via Valkey for logout/revocation
  version: 1.0.0
  contact:
    name: EMS Team
    email: support@ems.com
  license:
    name: Proprietary
    url: https://ems.com/license
servers:
  - url: http://localhost:8081
    description: Local development server
  - url: https://auth.ems.com
...
security:
  - bearerAuth: []
paths:
  /api/v1/auth/login:
    post:
      tags: [Authentication]
      summary: Login with email and password
      description: |
        Authenticate user using Direct Access Grant (Resource Owner Password Credentials).

        **Flow:**
        1. Client sends email/password
        2. Auth Facade exchanges credentials with IdP for tokens
        3. If MFA is enabled, returns `mfaRequired: true` with `mfaSessionToken`
        4. If successful, returns access token, refresh token, and user info

        **Rate Limiting:** 100 requests/minute per IP
      operationId: login
      security: []
      parameters:
        - $ref: '#/components/parameters/XTenantID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              standard:
                summary: Standard login
                value:
                  email: "user@example.com"
                  password: "securePassword123"
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
              examples:
                success:
                  summary: Successful login
                  value:
                    accessToken: "eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9..."
                    refreshToken: "eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9..."
                    expiresIn: 300
                    tokenType: "Bearer"
                    user:
                      id: "550e8400-e29b-41d4-a716-446655440000"
                      email: "user@example.com"
                      firstName: "John"
                      lastName: "Doe"
                      tenantId: "tenant-acme"
                      roles: ["ROLE_USER", "ROLE_MANAGER"]
                    mfaRequired: false
                mfaRequired:
                  summary: MFA required
                  value:
                    mfaRequired: true
                    mfaSessionToken: "mfa_session_abc123..."
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "invalid_credentials"
                message: "Invalid email or password"
                timestamp: "2024-01-15T10:30:00Z"
        '403':
          description: Account locked or disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "rate_limit_exceeded"
                message: "Too many login attempts. Please try again in 60 seconds."
                timestamp: "2024-01-15T10:30:00Z"
  /api/v1/auth/social/google:
    post:
      tags: [Social Login]
      summary: Login with Google One Tap
      description: |
        Exchange Google ID token for application tokens using RFC 8693 Token Exchange.

        **Frontend Integration:**
        ```javascript
        // Google One Tap callback
        function handleCredentialResponse(response) {
          fetch('/api/v1/auth/social/google', {
            method: 'POST',
            headers: {
              'Content-Type': 'application/json',
              'X-Tenant-ID': tenantId
            },
            body: JSON.stringify({ idToken: response.credential })
          });
        }
        ```

        **Token Exchange Flow:**
        1. Frontend obtains Google ID token via One Tap
        2. Auth Facade validates token with Google
        3. Token is exchanged with IdP (Keycloak) using `subject_issuer: google`
        4. Application tokens are returned
      operationId: loginWithGoogle
      security: []
      parameters:
        - $ref: '#/components/parameters/XTenantID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GoogleTokenRequest'
            example:
              idToken: "eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJhY2NvdW50cy5nb29nbGUuY29tIi..."
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid or expired Google token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "authentication_failed"
                message: "Google authentication failed"
                timestamp: "2024-01-15T10:30:00Z"
        '403':
          description: MFA required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /api/v1/auth/social/microsoft:
    post:
      tags: [Social Login]
      summary: Login with Microsoft MSAL
      description: |
        Exchange Microsoft access token for application tokens.

        **Frontend Integration (MSAL.js):**
        ```javascript
        const msalInstance = new msal.PublicClientApplication(msalConfig);
        const loginResponse = await msalInstance.loginPopup(loginRequest);

        fetch('/api/v1/auth/social/microsoft', {
          method: 'POST',
          headers: {
            'Content-Type': 'application/json',
            'X-Tenant-ID': tenantId
          },
          body: JSON.stringify({ accessToken: loginResponse.accessToken })
        });
        ```
      operationId: loginWithMicrosoft
      security: []
      parameters:
        - $ref: '#/components/parameters/XTenantID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MicrosoftTokenRequest'
            example:
              accessToken: "eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiIsIng1dCI6Ik5..."
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid or expired Microsoft token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: MFA required
  /api/v1/auth/login/{provider}:
    get:
      tags: [Dynamic Provider]
      summary: Initiate login with specific identity provider
      description: |
        Initiates OAuth2/OIDC login flow with a specific identity provider using
        Keycloak's `kc_idp_hint` mechanism (or equivalent for other providers).

        **Use Cases:**
        - Enterprise SSO (SAML/OIDC with corporate IdP)
        - Social login buttons that redirect rather than use tokens
        - B2B scenarios where tenants use their own IdP

        **Flow:**
        1. Client requests `/api/v1/auth/login/google?redirectUri=/administration`
        2. Auth Facade builds authorization URL with `kc_idp_hint=google`
        3. Returns HTTP 302 redirect to IdP login page
        4. User authenticates with IdP
        5. IdP redirects back to application with authorization code
        6. Application exchanges code for tokens

        **Supported Providers:**
        - `google` - Google OAuth2/OIDC
        - `microsoft` - Microsoft Azure AD
        - `facebook` - Facebook OAuth2
        - `github` - GitHub OAuth2
        - `saml` - Enterprise SAML (generic)
        - `okta` - Okta OIDC
        - `azure-ad` - Azure Active Directory
      operationId: initiateProviderLogin
      security: []
      parameters:
        - $ref: '#/components/parameters/XTenantID'
        - name: provider
          in: path
          required: true
          description: Identity provider alias (e.g., google, microsoft, saml)
          schema:
            type: string
            enum: [google, microsoft, facebook, github, saml, okta, azure-ad]
        - name: redirectUri
          in: query
          required: false
          description: URL to redirect after successful authentication
          schema:
            type: string
            default: "/"
      responses:
        '302':
          description: Redirect to identity provider
          headers:
            Location:
              description: IdP authorization URL
              schema:
                type: string
              example: "https://keycloak.example.com/realms/tenant-acme/protocol/openid-connect/auth?client_id=ems-auth-facade&redirect_uri=/administration&response_type=code&scope=openid%20profile%20email&kc_idp_hint=google&state=abc123"
        '200':
          description: Inline authentication successful (rare, for direct grant providers)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid provider or configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "invalid_provider"
                message: "Provider 'unknown' is not configured for this tenant"
                timestamp: "2024-01-15T10:30:00Z"
  /api/v1/auth/providers:
    get:
      tags: [Dynamic Provider]
      summary: List available identity providers
      description: |
        Returns the list of configured identity providers for the tenant.
        Used by frontend to display available login options.
      operationId: getAvailableProviders
      security: []
      parameters:
        - $ref: '#/components/parameters/XTenantID'
      responses:
        '200':
          description: Provider list retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProvidersResponse'
              example:
                active: "keycloak"
                tenant: "tenant-acme"
                available:
                  - alias: "google"
                    name: "Google"
                    type: "oidc"
                  - alias: "microsoft"
                    name: "Microsoft"
                    type: "oidc"
                  - alias: "saml"
                    name: "Enterprise SAML"
                    type: "saml"
...
  /api/v1/admin/tenants/{tenantId}/providers/{providerId}:
    get:
      tags: [Admin Provider Management]
      summary: Get identity provider by ID
      description: |
        Returns the configuration for a specific identity provider.
        Sensitive data like secrets are masked.

        **Requires:** ADMIN role
      operationId: getProvider
      parameters:
        - $ref: '#/components/parameters/TenantIdPath'
        - $ref: '#/components/parameters/ProviderIdPath'
      responses:
        '200':
          description: Provider retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderConfigResponse'
        '401':
          description: Not authenticated
        '403':
          description: Insufficient permissions
        '404':
          description: Provider not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags: [Admin Provider Management]
      summary: Update identity provider
      description: |
        Updates an existing identity provider configuration.

        **Note:** Changing the provider name or protocol may affect existing user sessions.

        **Requires:** ADMIN role
      operationId: updateProvider
      parameters:
        - $ref: '#/components/parameters/TenantIdPath'
        - $ref: '#/components/parameters/ProviderIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProviderConfigRequest'
      responses:
        '200':
          description: Provider updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderConfigResponse'
        '400':
          description: Invalid provider configuration
        '401':
          description: Not authenticated
        '403':
          description: Insufficient permissions
        '404':
          description: Provider not found

    delete:
      tags: [Admin Provider Management]
      summary: Delete identity provider
      description: |
        Deletes an identity provider from the tenant.

        **Warning:** This operation is irreversible. Users authenticated via
        this provider will need to re-authenticate using a different provider.

        **Requires:** ADMIN role
      operationId: deleteProvider
      parameters:
        - $ref: '#/components/parameters/TenantIdPath'
        - $ref: '#/components/parameters/ProviderIdPath'
      responses:
        '204':
          description: Provider deleted successfully
        '401':
          description: Not authenticated
        '403':
          description: Insufficient permissions
        '404':
          description: Provider not found
...
  /api/v1/auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh access token
      description: |
        Exchange refresh token for new access token with token rotation.

        **Token Rotation:**
        Each refresh request returns a new refresh token, invalidating the old one.
        This provides better security by limiting the window of opportunity for
        stolen tokens.

        **Best Practices:**
        - Store refresh token securely (httpOnly cookie or secure storage)
        - Refresh proactively before access token expires
        - Handle 401 by clearing tokens and redirecting to login
      operationId: refreshToken
      security: []
      parameters:
        - $ref: '#/components/parameters/XTenantID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
            example:
              refreshToken: "eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9..."
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "invalid_token"
                message: "Invalid or expired refresh token"
                timestamp: "2024-01-15T10:30:00Z"
  /api/v1/auth/logout:
    post:
      tags: [Authentication]
      summary: Logout
      description: |
        Invalidate refresh token and end the session.

        **What happens:**
        1. Refresh token is revoked at the IdP
        2. Token JTI is added to Valkey blacklist (TTL = token remaining lifetime)
        3. Client should clear local token storage

        **Note:** This is idempotent - calling logout with an already-revoked
        token returns success.
      operationId: logout
      security: []
      parameters:
        - $ref: '#/components/parameters/XTenantID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogoutRequest'
            example:
              refreshToken: "eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9..."
      responses:
        '204':
          description: Logout successful
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/mfa/setup:
    post:
      tags: [MFA]
      summary: Setup MFA
      description: |
        Initialize TOTP MFA setup for the authenticated user.

        **Setup Flow:**
        1. User requests MFA setup (this endpoint)
        2. Server generates TOTP secret and QR code
        3. User scans QR code with authenticator app
        4. User verifies setup with `/mfa/verify` endpoint
        5. MFA is confirmed and enabled

        **Response includes:**
        - `secret`: Base32-encoded TOTP secret (for manual entry)
        - `qrCodeUri`: Data URI for QR code image
        - `recoveryCodes`: One-time backup codes (store securely!)
      operationId: setupMfa
      parameters:
        - $ref: '#/components/parameters/XTenantID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MfaSetupRequest'
            example:
              method: "TOTP"
      responses:
        '200':
          description: MFA setup initialized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MfaSetupResponse'
              example:
                method: "TOTP"
                secret: "JBSWY3DPEHPK3PXP"
                qrCodeUri: "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAA..."
                recoveryCodes:
                  - "abc123-def456"
                  - "ghi789-jkl012"
                  - "mno345-pqr678"
                  - "stu901-vwx234"
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/mfa/verify:
    post:
      tags: [MFA]
      summary: Verify MFA code
      description: |
        Verify TOTP code to complete MFA authentication or setup.

        **Two use cases:**

        1. **During Login (MFA Challenge):**
           - After login returns `mfaRequired: true`
           - User enters TOTP code from authenticator
           - On success, returns full auth tokens

        2. **During MFA Setup:**
           - After `/mfa/setup` returns secret and QR code
           - User enters code to confirm setup
           - On success, MFA is permanently enabled

        **Code Validation:**
        - 6-digit TOTP code
        - 30-second time window with 1-step tolerance
        - SHA1 hash algorithm (standard)
      operationId: verifyMfa
      security: []
      parameters:
        - $ref: '#/components/parameters/XTenantID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MfaVerifyRequest'
            example:
              mfaSessionToken: "mfa_session_abc123..."
              code: "123456"
      responses:
        '200':
          description: MFA verified successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid MFA code or session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "invalid_mfa_code"
                message: "Invalid MFA code"
                timestamp: "2024-01-15T10:30:00Z"
  /api/v1/auth/me:
    get:
      tags: [User Profile]
      summary: Get current user profile
      description: |
        Get profile information for the authenticated user.

        User information is extracted from the JWT claims using the
        provider-agnostic claim mapping configured in `auth.facade.user-claim-mappings`.
      operationId: getCurrentUser
      responses:
        '200':
          description: User profile retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfo'
              example:
                id: "550e8400-e29b-41d4-a716-446655440000"
                email: "user@example.com"
                firstName: "John"
                lastName: "Doe"
                tenantId: "tenant-acme"
                roles:
                  - "ROLE_USER"
                  - "ROLE_MANAGER"
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "unauthorized"
                message: "Not authenticated"
                timestamp: "2024-01-15T10:30:00Z"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token obtained from login endpoints.

        **Header format:** `Authorization: Bearer <token>`

        **Token structure:**
        - `sub`: User ID
        - `email`: User email
        - `tenant_id`: Tenant identifier
        - `realm_access.roles`: User roles (Keycloak)
        - `exp`: Expiration timestamp
  parameters:
    XTenantID:
      name: X-Tenant-ID
      in: header
      required: true
      description: |
        Tenant identifier. Maps to IdP realm/organization.

        **Format:** `tenant-{name}` or `master` for platform admin

        **Examples:**
        - `tenant-acme` - ACME Corporation tenant
        - `tenant-globex` - Globex tenant
        - `master` - Platform administration
      schema:
        type: string
        pattern: '^(master|tenant-[a-z0-9-]+)$'
      example: "tenant-acme"
...